+++ extensions-core/druid-kerberos/src/main/resources/META-INF/services/io.druid.initialization.DruidModule
@@ -1,1 +1,1 @@
-io.druid.kerberos.DruidKerberosModule
+io.druid.security.kerberos.DruidKerberosModule
