+++ extensions-core/druid-kerberos/src/test/java/io/druid/security/kerberos/DruidKerberosConfigTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.druid.kerberos;
+package io.druid.security.kerberos;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Binder;
